requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
anthropic==0.40.0
openai==1.57.0
openpyxl==3.1.5
pandas==2.2.0
Flask==3.0.0
Werkzeug==3.0.1
